health-phenotype-discovery/requirements.txt
# Dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
seaborn>=0.12.0
xgboost>=1.7.0
shap>=0.41.0
pyyaml>=6.0
requests>=2.28.0
jupyter>=1.0.0
notebook>=6.5.0
